Fixes for new behaviour in art sute v3_15_00.
Includes that never belonged in .hh files caused the dictionary maker to fail.
If you follow the include chains on the offending includes, it eventually
includes files from fhiclcpp, which now have content that the dictionary
maker does not know how to deal with.  These changes are of two types:

    Remove superfluous includes.

    Move some includes from .hh files to .cc files.